backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary,pool]>=3.1.0",
    "langgraph>=0.6.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-openai>=0.1.0",
    "langchain>=0.2.0",
    "greenlet>=3.2.4",
    "scipy>=1.16.1",
    "json-repair>=0.48.0",
]
